Fix signed/unsigned comparison compiler warning

+++ src/analyzer/protocol/udp/UDP.cc
@@ -67,7 +67,7 @@ void UDP_Analyzer::DeliverPacket(int len, const u_char* data, bool is_orig,
 	constexpr auto eth_len = 14;
 
 	if ( validate_checksum &&
-	     len > (sizeof(struct udphdr) + vxlan_len + eth_len) &&
+	     len > ((int)sizeof(struct udphdr) + vxlan_len + eth_len) &&
 	     (data[0] & 0x08) == 0x08 )
 		{
 		auto& vxlan_ports = analyzer_mgr->GetVxlanPorts();
